+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## 2.8.13 - *tbd*
+
+### Bugfixes
+
+ - *tbd*
+
+### Changes
+
+ - *tbd*
+
+### Added
+
+ - *tbd*
+
+### Removed
+
+ - *tbd*
+
+### Internal Improvements and Changes
+
+ - *tbd*
+
+
 ## 2.8.12 - Coridian (2021-05-10)
 
 ### Bugfixes